moved auth-ui to own component, added redirect
