+++ src/app/login/page.tsx
@@ -1,34 +1,30 @@
-"use client"
+import { createClient } from "@/utils/supabase/server"
+import { cookies } from "next/headers"
+import { redirect } from "next/navigation"
 
-import { getUrl } from "@/utils/misc"
-import { createClient } from "@/utils/supabase/client"
-import { Auth } from "@supabase/auth-ui-react"
-import { ThemeSupa } from "@supabase/auth-ui-shared"
+import { Card, CardContent } from "@/components/ui/card"
 
-export default function LoginPage() {
-  const supabase = createClient()
+import AuthUI from "./auth-ui"
+
+export default async function LoginPage() {
+  const supabase = createClient(cookies())
+
+  const {
+    data: { session },
+  } = await supabase.auth.getSession()
+
+  if (session) {
+    return redirect("/")
+  }
 
   return (
     <main className="container py-16">
       <div className="mx-auto max-w-md">
-        <Auth
-          supabaseClient={supabase}
-          appearance={{
-            theme: ThemeSupa,
-            variables: {
-              default: {
-                colors: {
-                  brand: "#6366f1",
-                  brandAccent: "#6366f1",
-                },
-              },
-            },
-          }}
-          providers={["github"]}
-          view="magic_link"
-          showLinks={false}
-          redirectTo={`${getUrl()}/auth/callback`}
-        />
+        <Card>
+          <CardContent className="pt-3">
+            <AuthUI />
+          </CardContent>
+        </Card>
       </div>
     </main>
   )
